fixed errors, added better comments, added a dashboard for users logged in

+++ controllers/api/comment-routes.js
@@ -1,6 +1,7 @@
 const router = require("express").Router();
 const { Comment } = require("../../models");
 
+// Get all comments
 router.get("/", (req, res) => {
   Comment.findAll()
     .then((dbCommentData) => res.json(dbCommentData))
@@ -10,14 +11,13 @@ router.get("/", (req, res) => {
     });
 });
 
+// Create a new comment
 router.post("/", (req, res) => {
-  // check the session
   if (req.session) {
     Comment.create({
       comment_text: req.body.comment_text,
-      post_id: req.body.post_id,
-      // use the id from the session
       user_id: req.session.user_id,
+      post_id: req.body.post_id,
     })
       .then((dbCommentData) => res.json(dbCommentData))
       .catch((err) => {
@@ -27,23 +27,26 @@ router.post("/", (req, res) => {
   }
 });
 
+// Delete a comment by id
 router.delete("/:id", (req, res) => {
-  Comment.destroy({
-    where: {
-      id: req.params.id,
-    },
-  })
-    .then((dbCommentData) => {
-      if (!dbCommentData) {
-        res.status(404).json({ message: "No comment found with this id!" });
-        return;
-      }
-      res.json(dbCommentData);
+  if (req.session) {
+    Comment.destroy({
+      where: {
+        id: req.params.id,
+      },
     })
-    .catch((err) => {
-      console.log(err);
-      res.status(500).json(err);
-    });
+      .then((dbCommentData) => {
+        if (!dbCommentData) {
+          res.status(404).json({ message: "No comment found with this id!" });
+          return;
+        }
+        res.json(dbCommentData);
+      })
+      .catch((err) => {
+        console.log(err);
+        res.status(500).json(err);
+      });
+  }
 });
 
 module.exports = router;
